More robust label serializer with shape based on InputImage
